% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collate_all_excel}
\alias{collate_all_excel}
\title{Collect all Excel files}
\usage{
collate_all_excel(dir, dest = NULL)
}
\arguments{
\item{dir}{Root directory to commence search for files.}

\item{dest}{Directory where collated files are stored.}
}
\description{
Collects all the Excel files within a given parent directory and its
children.
}
