% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collate_excel}
\alias{collate_excel}
\title{Collect all Excel files}
\usage{
collate_excel(dir, dest = NULL, verbose = TRUE)
}
\arguments{
\item{dir}{Root directory to commence search for files.}

\item{dest}{Directory where collated files are stored.}

\item{verbose}{Logical. Whether to display a printout of copy operations.}
}
\description{
Collects all the Excel files within a given parent directory and its
children.
}
