% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine_dfs}
\alias{combine_dfs}
\title{Combines all the data frames in the list into one master data frame}
\usage{
combine_dfs(dfs)
}
\arguments{
\item{dfs}{A list of data frames}
}
\value{
A combined data frame from all previously existing one.
}
\description{
Combines all the data frames in the list into one master data frame
}
