% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_phone_numbers}
\alias{fix_phone_numbers}
\title{Fixing up phone numbers to a uniform text format}
\usage{
fix_phone_numbers(column)
}
\arguments{
\item{column}{A column with mobile phone mumbers (Nigeria format)}
}
\value{
A vector of updated phone numbers
}
\description{
Fixing up phone numbers to a uniform text format
}
