% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gather_excel_files}
\alias{gather_excel_files}
\title{Collect Excel Files}
\usage{
gather_excel_files()
}
\value{
The function has no return value.
}
\description{
Searches the machine's HOME directory tree for Excel files and puts them
into a separate directory.
}
\details{
This function is a wrapper for \code{\link[exhale]{collate_excel}}
}
