% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise_narc_excel}
\alias{harmonise_narc_excel}
\title{Harmonise Nehemiah's Data}
\usage{
harmonise_narc_excel(dir)
}
\arguments{
\item{dir}{Path to the folder (hopefully) containing Excel files}
}
\description{
Merge data belonging to Nehemia Apostolic Resource Centre into one database
file.
}
