% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{locate_header}
\alias{locate_header}
\title{Helper Functions.}
\usage{
locate_header(df, hdr, quietly = TRUE)
}
\arguments{
\item{df}{A single data frame from NARC spreeadsheet}

\item{hdr}{The new header for the final data frame}

\item{quietly}{logical - Whether to print out progress or not}
}
\value{
An S3 object that is a marker to the row it occupies
}
\description{
Helper Functions.
}
\details{
Finds the row that likely contains the actual header
}
