% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narc_harmonise.R
\name{narc_harmonise}
\alias{narc_harmonise}
\title{Harmonise NARC Data From Different Excel Spreadsheets.}
\usage{
narc_harmonise(path = NULL)
}
\arguments{
\item{path}{Path to the folder (hopefully) containing Excel files}
}
\description{
\code{narc_harmonise}
}
