% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rearrange_df}
\alias{rearrange_df}
\title{Rearranges the columns of data frames to suit the prescribed
format. Columns without values are assigned NA.}
\usage{
rearrange_df(df, hdr)
}
\arguments{
\item{df}{A data frame taken from NARC Excel spreadsheet}

\item{hdr}{A template header used by NARC (see 'Details')}
}
\value{
A data frame with rearranged headers and columns
}
\description{
Rearranges the columns of data frames to suit the prescribed
format. Columns without values are assigned NA.
}
