% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_datatypes}
\alias{set_datatypes}
\title{Setting the columns to the appropriate data types}
\usage{
set_datatypes(df)
}
\arguments{
\item{df}{A data frame of NARC mailing list data}
}
\value{
A dataframe with updated datatypes
}
\description{
Setting the columns to the appropriate data types
}
