% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_header}
\alias{update_header}
\title{Updates old column names of the data frame with the new
ones that are to be used in the harmonised version}
\usage{
update_header(df, newCol)
}
\arguments{
\item{df}{A data frame taken from NARC mailing list spreadsheet}

\item{newCol}{The proposed new format for the final data frame}
}
\value{
A data frame with updated header
}
\description{
Updates old column names of the data frame with the new
ones that are to be used in the harmonised version
}
