% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate_narc_mail}
\alias{consolidate_narc_mail}
\title{Consolidation of Database entries (with focus on duplications)}
\usage{
consolidate_narc_mail(db, table = "NARC_mail")
}
\arguments{
\item{db}{The database file (SQLite).}

\item{table}{The database table to be consolidated; a character vector of
length 1.}
}
\value{
The functions returns a copy of the consolidated data frame. A side
effect is the creation of a table through internal implementation. That
table is named \emph{consolidated}.
}
\description{
Consolidation of Database entries (with focus on duplications)
}
\details{
The default argument for the \code{table} parameter is
\emph{NARC_mail}. This is kept for backward compatibility with earlier
code that was specifically tailored for use by this function.
}
\note{
In its current state, the function preloads the data internally.
Thus, regardless of the way data are loaded in the global environment,
they are reloaded from the table from scratch. Where the table does not
have a 'serialno' column, the function fails. The real problem is the fact
that the function is tailor-made for a data table that comes with specific
field format. This is something that would be reconsidered in future.
}
