% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate_narc_mail}
\alias{consolidate_narc_mail}
\title{Consolidation of Database entries (with focus on duplications)}
\usage{
consolidate_narc_mail(db)
}
\description{
Consolidation of Database entries (with focus on duplications)
}
