% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rearrange_df}
\alias{rearrange_df}
\title{Rearrange data frame columns}
\usage{
rearrange_df(df, hdr)
}
\arguments{
\item{df}{An object of class \code{data.frame}}

\item{hdr}{A character vector containing names of outputted data frame
appearing in the desired order.}
}
\value{
A data frame.
}
\description{
Rearranges the columns of data frames to suit the prescribed
format. Columns without values are assigned NA.
}
